Beacon and device have current user set as owner

+++ src/main/java/hr/foi/diplomski/central/service/beacon/BeaconServiceImpl.java
@@ -5,10 +5,11 @@ import hr.foi.diplomski.central.controllers.api.beacons.dto.BeaconViewDto;
 import hr.foi.diplomski.central.exceptions.BadRequestException;
 import hr.foi.diplomski.central.mappers.beacon.BeaconMapper;
 import hr.foi.diplomski.central.model.Beacon;
-import hr.foi.diplomski.central.model.Device;
+import hr.foi.diplomski.central.model.User;
 import hr.foi.diplomski.central.repository.BeaconRepository;
 import hr.foi.diplomski.central.repository.DeviceRepository;
 import hr.foi.diplomski.central.repository.RecordRepository;
+import hr.foi.diplomski.central.service.users.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -24,10 +25,11 @@ public class BeaconServiceImpl implements BeaconService {
     private final DeviceRepository deviceRepository;
     private final RecordRepository recordRepository;
     private final BeaconMapper beaconMapper;
+    private final UserService userService;
 
     @Override
     public List<BeaconDto> findAllBeacons() {
-        return beaconMapper.entitysToDto(beaconRepository.findAll());
+        return beaconMapper.entitysToDto(beaconRepository.findAllByUser(userService.findCurrentUser()));
     }
 
     @Override
@@ -35,14 +37,33 @@ public class BeaconServiceImpl implements BeaconService {
         Beacon beacon = beaconRepository.findById(id).orElseThrow(() -> new BadRequestException(
                 String.format("Ne postoji beacon s id: %s", id)));
 
-        return beaconMapper.entityToDto(beacon);
+        User user = userService.findCurrentUser();
+
+        if (user.equals(beacon.getUser())) {
+            return beaconMapper.entityToDto(beacon);
+        } else {
+            throw new BadRequestException(String.format("Korisnik %s nije vlasnik beacona %s i ne može ga pregledati",
+                    user.getUsername(), beacon.getId()));
+        }
+
     }
 
     @Override
     public BeaconDto saveBeacon(BeaconDto beaconDto) {
         Beacon beacon = beaconMapper.dtoToEntity(beaconDto);
+        User user = userService.findCurrentUser();
 
-        beacon = beaconRepository.save(beacon);
+        if (beacon.getId() != null && beacon.getId() != 0) {
+            if (user.equals(beacon.getUser())) {
+                beacon = beaconRepository.save(beacon);
+            } else {
+                throw new BadRequestException(String.format("Korisnik %s nije vlasnik beacona %s i ne može ga uređivati",
+                        user.getUsername(), beacon.getId()));
+            }
+        } else {
+            beacon.setUser(user);
+            beacon = beaconRepository.save(beacon);
+        }
 
         return beaconMapper.entityToDto(beacon);
     }
@@ -53,16 +74,23 @@ public class BeaconServiceImpl implements BeaconService {
                 .orElseThrow(() -> new BadRequestException(String
                         .format("Ne postoji beacon s id: %s", idBeacon)));
 
-        recordRepository.deleteAllByRecordId_Beacon(beacon);
+        User user = userService.findCurrentUser();
 
-        beacon.getDevices().forEach(e -> e.setBeacon(null));
-        beacon.getDevices().clear();
+        if (user.equals(beacon.getUser())) {
+            recordRepository.deleteAllByRecordId_Beacon(beacon);
 
-        beaconRepository.delete(beacon);
+            beacon.getDevices().forEach(e -> e.setBeacon(null));
+            beacon.getDevices().clear();
+
+            beaconRepository.delete(beacon);
+        } else {
+            throw new BadRequestException(String.format("Korisnik %s nije vlasnik beacona %s i ne može ga obrisati",
+                    user.getUsername(), beacon.getId()));
+        }
     }
 
     @Override
     public List<BeaconViewDto> findAllFreeBeacons() {
-        return beaconMapper.entitysToViewDtos(beaconRepository.findAllByDevicesEmpty());
+        return beaconMapper.entitysToViewDtos(beaconRepository.findAllByUserAndDevicesIsEmpty(userService.findCurrentUser()));
     }
 }
